refactor: Tabs 컴포넌트를 shared로 이동
