+++ frontend/src/components/@shared/Tabs/type.ts
