+++ frontend/src/components/index.ts
@@ -17,6 +17,7 @@ export * from 'components/@shared/ShareButton';
 export * from 'components/@shared/Tooltip';
 export * from 'components/@shared/ToggleButton';
 export * from 'components/@shared/UnderLineText';
+export * from 'components/@shared/Tabs';
 export * from 'components/Logo';
 export * from 'components/ValidationMessage';
 export * from 'components/Header';
@@ -51,5 +52,3 @@ export * from 'components/ErrorFallbackSubscriptionButton';
 export * from 'components/ErrorFallbackNotificationMessage';
 export * from 'components/MembersPopover';
 export * from 'components/CheckSuccessCycle';
-export * from 'components/Tabs';
-export * from 'components/SROnly';
